iio: meter: ADE9078 fixed nested interrupts
iio_trigger_poll is interpreted as an interrupt and thus by calling this
function inside any interrupt handler will lead to a nested interrupt
scenario which can be avoided.

In order to avoid nesting a ~10ms rudimentary timeout has been
implemented which checks for the trigger conditions. If they are met
all irqs are sored and disabled before the iio_trigger_poll is called,
at the end of which they are restored and re-enabled